pyproject.toml
[project]
name = "stock-selector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.24.1",
    "langchain-community>=0.3.27",
    "langgraph>=0.6.3",
    "langsmith>=0.4.12",
    "python-dotenv>=1.1.1",
]
